Revert "fix security issue"
This reverts commit 4bf7f46a167cfc9607709a25a09720620fe02e84.

+++ multi_physics/QED/QED_tests/test_picsar_tables.cpp
@@ -387,8 +387,7 @@ BOOST_AUTO_TEST_CASE( picsar_equispaced_2d_table_setall)
     auto tab_2d = make_2d_table();
     const auto val = 42.0;
     const auto vec = std::vector<double>(
-        static_cast<std::size_t>(tab_2d.get_how_many_x())*
-        static_cast<std::size_t>(tab_2d.get_how_many_y()),
+        tab_2d.get_how_many_x()*tab_2d.get_how_many_y(),
         val);
 
     tab_2d.set_all_vals(vec);
